Cleanup of app.py dependency mess
