+++ foldatlas/controllers.py
@@ -1,7 +1,5 @@
-from database import db_session
 from sqlalchemy import and_
 
-import json, database, settings, uuid, os, subprocess
 from models import Feature, Transcript, NucleotideMeasurementSet, Structure, \
     GeneLocation, NucleotideMeasurementRun, StructurePredictionRun, \
     values_str_unpack_float, values_str_unpack_int, RawReactivities
@@ -14,6 +12,7 @@ from utils import ensure_dir, insert_newlines, build_dot_bracket
 class GenomeBrowser():
 
     def get_transcripts(self, request):
+        import database
 
         chromosome_id = "Chr"+str(int(request.args.get('chr'))) # SQL-injection safe
         start = int(request.args.get('start'))
@@ -84,6 +83,7 @@ class GenomeBrowser():
     def get_genes(self, request):
         
         from utils import Timeline
+        import database
 
         chromosome_id = "Chr"+str(int(request.args.get('chr'))) # SQL-injection safe
         start = int(request.args.get('start'))
@@ -113,6 +113,7 @@ class GenomeBrowser():
 
     # Fetch chromosome IDs and their lengths. Used for chromosome menu and also initialising the genome browser.
     def get_chromosomes(self):
+        import database
 
         sql = ( "SELECT chromosome_id, CHAR_LENGTH(sequence) length FROM chromosome "
                 "WHERE strain_id = '"+settings.reference_strain_id+"' "
@@ -133,6 +134,7 @@ class GenomeBrowser():
 class TranscriptView():
 
     def __init__(self, transcript_id):
+        from database import db_session
 
         self.transcript_id = transcript_id
 
@@ -174,6 +176,7 @@ class NucleotideMeasurementView():
     def build_entries(self, experiment_ids):
 
         from models import NucleotideMeasurementRun
+        from database import db_session
 
         # Load experiments
         experiments = db_session \
@@ -247,6 +250,8 @@ class AlignmentView():
         self.build_alignment_entries()
 
     def build_alignment_entries(self):
+        from database import db_session
+
         self.alignment_rows = []
 
         # fetch the alignment rows from the DB, using the ORM
@@ -317,6 +322,7 @@ class AlignmentView():
 class TranscriptSearcher():
     def search(self, search_string):
         from flask import abort
+        from database import db_session
 
         transcripts = db_session \
             .query(Transcript) \
@@ -346,6 +352,7 @@ class CoverageSearcher():
         # better to do the imports closer to where they are needed
         from sqlalchemy import func
         from math import ceil
+        from database import db_session
 
         transcript_count = db_session \
             .query(func.count('*')) \
@@ -360,6 +367,7 @@ class CoverageSearcher():
         from utils import Timeline
         from sqlalchemy import func, and_
         from models import Structure, GeneLocation
+        import database
 
         offset = (int(page_num) - 1) * self.page_size
         limit = self.page_size
@@ -394,7 +402,7 @@ class CoverageSearcher():
             "GROUP BY jnms.transcript_id "
             "ORDER BY coverage DESC"
         )
-
+    
         results = database.engine.execute(sql)
 
         out = []
@@ -472,6 +480,7 @@ class StructureView():
     def build_entries(self, structure_prediction_run_ids):
 
         from models import Structure, StructurePredictionRun
+        from database import db_session
 
         # Load experiments
         runs = db_session \
@@ -535,6 +544,7 @@ class StructureDiagramView():
         self.data_json = json.dumps(data)
 
     def build_dot_bracket(self):
+        from database import db_session
 
         # get all the positions
         results = db_session \
@@ -598,6 +608,9 @@ class StructureCirclePlotView():
         self.get_values()
 
     def get_values(self):
+
+        from database import db_session
+
         # convert entities to dot bracket string
         out = [];
 
@@ -635,6 +648,7 @@ class StructureDownloader():
         self.transcript_id = transcript_id
 
     def generate(self):
+        from database import db_session
 
         # Fetch the data
         results = db_session \
@@ -715,6 +729,8 @@ class NucleotideMeasurementDownloader():
 
     # Retrieves raw reactivity values and outputs as text
     def get_raw(self):
+        from database import db_session
+
         seq_str = Transcript(self.transcript_id).get_sequence_str()
 
         # Use the ORM to grab raw data
@@ -744,6 +760,8 @@ class NucleotideMeasurementDownloader():
 
     # Retrieves normalised reactivities and outputs as text
     def get_normalised(self):
+
+        from database import db_session
 
         # Grab sequence string
         seq_str = Transcript(self.transcript_id).get_sequence_str()
